octotunnel/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "octotunnel"
version = "0.1.0"
description = "Spin up pod-like network namespaces with independent AnyConnect/OpenConnect sessions."
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
  "PyYAML>=6.0",
]

[project.scripts]
octotunnel = "octotunnel.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
